Rename histograming -> histogramming.
