+++ analysis_templates/cms_minimal/__cf_module_name__/histogramming/example.py
@@ -4,8 +4,8 @@
 Example histogram producer.
 """
 
-from columnflow.histograming import HistProducer
-from columnflow.histograming.default import cf_default
+from columnflow.histogramming import HistProducer
+from columnflow.histogramming.default import cf_default
 from columnflow.util import maybe_import
 from columnflow.config_util import get_shifts_from_sources
 from columnflow.columnar_util import Route
