+++ analysis_templates/cms_minimal/law.cfg
@@ -32,7 +32,7 @@ selection_modules: columnflow.selection.empty, columnflow.selection.cms.{json_fi
 reduction_modules: columnflow.reduction.default, __cf_module_name__.reduction.example
 production_modules: columnflow.production.{categories,matching,normalization,processes}, columnflow.production.cms.{btag,electron,jet,matching,mc_weight,muon,pdf,pileup,scale,parton_shower,seeds}, __cf_module_name__.production.example
 categorization_modules: __cf_module_name__.categorization.example
-hist_production_modules: columnflow.histograming.default, __cf_module_name__.histograming.example
+hist_production_modules: columnflow.histogramming.default, __cf_module_name__.histogramming.example
 ml_modules: columnflow.ml, __cf_module_name__.ml.example
 inference_modules: columnflow.inference, __cf_module_name__.inference.example
