--- docs/api/histograming/default.rst
@@ -1,9 +1,0 @@
-``default``
-===========
-
-.. currentmodule:: columnflow.histograming.default
-
-.. automodule:: columnflow.histograming.default
-    :autosummary:
-    :members:
-    :undoc-members:
