--- docs/api/histograming/index.rst
@@ -1,14 +1,0 @@
-``columnflow.histograming``
-===========================
-
-.. currentmodule:: columnflow.histograming
-
-.. automodule:: columnflow.histograming
-    :autosummary:
-    :members:
-    :undoc-members:
-
-.. toctree::
-    :maxdepth: 1
-
-    default
