+++ docs/api/histogramming/default.rst
@@ -1,0 +1,9 @@
+``default``
+===========
+
+.. currentmodule:: columnflow.histogramming.default
+
+.. automodule:: columnflow.histogramming.default
+    :autosummary:
+    :members:
+    :undoc-members:
