+++ docs/api/histogramming/index.rst
@@ -1,0 +1,14 @@
+``columnflow.histogramming``
+============================
+
+.. currentmodule:: columnflow.histogramming
+
+.. automodule:: columnflow.histogramming
+    :autosummary:
+    :members:
+    :undoc-members:
+
+.. toctree::
+    :maxdepth: 1
+
+    default
