+++ law.cfg
@@ -26,7 +26,7 @@ calibration_modules: columnflow.calibration
 selection_modules: columnflow.selection.empty
 reduction_modules: columnflow.reduction.default
 categorization_modules: columnflow.categorization
-hist_production_modules: columnflow.histograming.default
+hist_production_modules: columnflow.histogramming.default
 ml_modules: columnflow.ml
 inference_modules: columnflow.inference
